pcall: port to lua-5.4
lua 5.4 changes the `lua_resume` API:

```
The function lua_resume has an extra parameter. This out parameter
returns the number of values on the top of the stack that were yielded
or returned by the coroutine. (In previous versions, those values were
the entire stack.)
```

https://www.lua.org/manual/5.4/manual.html#8.3

this patch is similar to the implementation in the sol bindings